Removed compiler warnings for DParser.
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9873 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/dparser/symtab.c
@@ -91,7 +91,7 @@ new_D_SymHash() {
 
 static void
 free_D_SymHash(D_SymHash *sh) {
-  int i;
+  uint i;
   D_Sym *sym;
   for (i = 0; i < sh->syms.n; i++)
     for (; sh->syms.v[i]; sh->syms.v[i] = sym) {
@@ -267,7 +267,7 @@ commit_ll(D_Scope *st, D_SymHash *sh) {
 /* make direct links to the latest update */
 static void 
 commit_update(D_Scope *st, D_SymHash *sh) {
-  int i;
+  uint i;
   D_Sym *s;
 
   for (i = 0; i < sh->syms.n; i++)
@@ -289,7 +289,7 @@ commit_D_Scope(D_Scope *st) {
 
 D_Sym *
 new_D_Sym(D_Scope *st, char *name, char *end, int sizeof_D_Sym) {
-  int len = end ? end - name : name ? strlen(name) : 0;
+  uint len = end ? (uint)(end - name) : name ? strlen(name) : 0;
   D_Sym *s = MALLOC(sizeof_D_Sym);
   memset(s, 0, sizeof_D_Sym);
   s->name = name;
@@ -376,7 +376,7 @@ find_D_Sym_internal(D_Scope *cur, char *name, int len, uint h) {
 
 D_Sym *
 find_D_Sym(D_Scope *st, char *name, char *end) {
-  int len = end ? end - name : strlen(name);
+  uint len = end ? (uint)(end - name) : strlen(name);
   uint h = strhashl(name, len);
   D_Sym *s = find_D_Sym_internal(st, name, len, h);
   if (s)
@@ -387,7 +387,7 @@ find_D_Sym(D_Scope *st, char *name, char *end) {
 D_Sym *
 find_global_D_Sym(D_Scope *st, char *name, char *end) {
   D_Sym *s;
-  int len = end ? end - name : strlen(name);
+  uint len = end ? (uint)(end - name) : strlen(name);
   uint h = strhashl(name, len);
   D_Scope *cur = st;
   while (cur->up) cur = cur->search;
@@ -399,7 +399,7 @@ find_global_D_Sym(D_Scope *st, char *name, char *end) {
 
 D_Sym *
 find_D_Sym_in_Scope(D_Scope *st, D_Scope *cur, char *name, char *end) {
-  int len = end ? end - name : strlen(name);
+  uint len = end ? (uint)(end - name) : strlen(name);
   uint h = strhashl(name, len);
   D_Sym *s = find_D_Sym_in_Scope_internal(cur, name, len, h);
   if (s)
@@ -418,7 +418,7 @@ next_D_Sym_in_Scope(D_Scope **scope, D_Sym **sym) {
   }
   for (;st ; st = st->search) {
     if (st->hash) {
-      int i = last_sym ? ((last_sym->hash + 1) % st->hash->syms.n) : 0;
+      uint i = last_sym ? ((last_sym->hash + 1) % st->hash->syms.n) : 0;
       if (!last_sym || i)
 	ll = st->hash->syms.v[i];
     } else {
@@ -475,7 +475,7 @@ print_scope(D_Scope *st) {
   if (st->ll) printf("  LL\n");
   if (st->hash) printf("  HASH\n");
   if (st->hash) {
-    int i;
+    uint i;
     for (i = 0; i < st->hash->syms.n; i++)
       if (st->hash->syms.v[i])
 	print_sym(st->hash->syms.v[i]);
